2018/9/6

+++ src/test/java/com/team6/dao/UserMapperTest.java
@@ -17,12 +17,13 @@ public class UserMapperTest {
     private UserMapper userMapper;
     @Test
     public void queryById() {
-        long id = 10001;
+        int id = 10001;
         User user = userMapper.queryById(id);
         System.out.println(user);
     }
 
     @Test
+
     public void queryByName() {
         String name = "郑伟坑";
         User user = userMapper.queryByName(name);
